+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Allow steps to be placed relative to other steps using `before_step:` and `after_step:` keyword arguments
 - Use just-in-time index lookup instead of `Step#seq`
 
 ## [0.1.18] - 2025-06-20